Moved LogNode to Node
