+++ shared/node.go
@@ -1,5 +1,9 @@
 package shared
 
+import (
+	"time"
+)
+
 // ParserNode contains general node data
 type ParserNode struct {
 	Header   string
@@ -20,6 +24,7 @@ type DBNode struct {
 	Elements Elements
 }
 
+// NewDBNodeFromNode creates DB Node from Parser node
 func NewDBNodeFromNode(n *ParserNode) *DBNode {
 	dbn := DBNode(*n)
 	return &dbn
@@ -30,11 +35,41 @@ type DBNodeList map[string]*DBNode
 
 // NewDBNodeList creates new list of general nodes
 func NewDBNodeList() DBNodeList {
-	dbnl := &DBNodeList{}
-	return *dbnl
+	return DBNodeList{}
 }
 
 // Push adds node to the node list
 func (db *DBNodeList) Push(node *DBNode) {
 	(*db)[(*node).Header] = node
+}
+
+// LogNode contains log node data
+type LogNode struct {
+	Time     time.Time
+	Elements Elements
+}
+
+// NewLogNode creates new log node
+func NewLogNode(time time.Time, elements Elements) *LogNode {
+	return &LogNode{time, elements}
+}
+
+// NewLogNodeFromNode creates new LogNode from ParserNode
+func NewLogNodeFromNode(node *ParserNode, dateFormat string) (*LogNode, error) {
+	t, err := time.Parse(dateFormat, node.Header)
+	if err != nil {
+		return nil, err
+	}
+
+	elList := NewElements()
+
+	for _, el := range node.Elements {
+		if ndx, exists := elList.Index(el.Name); exists {
+			elList[ndx].Val += el.Val
+		} else {
+			elList.Add(el.Name, el.Val)
+		}
+	}
+
+	return NewLogNode(t, elList), nil
 }
